brats-segmentation/requirements.txt
torch>=2.1.0
monai[all]>=1.3.0
nnunetv2>=2.5
nibabel>=5.0
numpy>=1.24
matplotlib>=3.7
scipy>=1.10
pandas>=2.0
tqdm>=4.65
rich>=13.0
pyyaml>=6.0
tensorboard>=2.14
scikit-learn>=1.3
seaborn>=0.12
